add energy value to float

+++ backend/src/clients/database/models/price.py
@@ -11,10 +11,10 @@ class Price(Base):
     size_id: Mapped[int] = mapped_column(ForeignKey("sizes.size_id", ondelete="CASCADE"))
     product_id: Mapped[int] = mapped_column(ForeignKey("products.product_id", ondelete="CASCADE"))
     price: Mapped[float] = mapped_column(nullable=False)
-    proteins: Mapped[int] = mapped_column(nullable=True)
-    fats: Mapped[int] = mapped_column(nullable=True)
-    carbohydrates: Mapped[int] = mapped_column(nullable=True)
-    calories: Mapped[int] = mapped_column(nullable=True)
+    proteins: Mapped[float] = mapped_column(nullable=True)
+    fats: Mapped[float] = mapped_column(nullable=True)
+    carbohydrates: Mapped[float] = mapped_column(nullable=True)
+    calories: Mapped[float] = mapped_column(nullable=True)
     is_custom: Mapped[bool] = mapped_column(nullable=False)
 
     product: Mapped["Product"] = relationship(back_populates="prices")  # noqa: F821
